last backup in schedule determines backup (fixes #1)
Read the configuration file and the backup location to
determine if each period for the backup is required or not.

This is still slightly buggy because requiring a backup for a
longer period should trigger the need for backup in all
shorter periods. I.e. if weekly is needed, then daily and
hourly are needed, too, if active in the configuration.

Signed-off-by: Eddy Petrișor <eddy.petrisor@gmail.com>