bugfix: avoid race condition when ran from cron
When ran va cron it is possible that we end up in a race condition
and we run 2 or more occasional_rsnapshot instances at a time.

Avoid the situation where two consecutive executions via cron
synchronize in such a way that the later spawned
occasional_rsnapshot inserts itself right between a larger interval
and the lower next one from the first's execution.

This can happen when using delayed removal of the oldest snapshot
because the rsnapshot pid is removed at this point; problem is that
occasional_rsnapshot hasn't finished at this point

We avoid this situation by making our own lockfile and check it
before doing anything else.

Signed-off-by: Eddy Petrișor <eddy.petrisor@gmail.com>