bugfix: avoid repeated backups for unpopulated intervals
When having only a few backups in the shorter intervals, all the
longer intervals were considered candidates for backup. That was
stupid since for new backup locations one would not have any of the
longer intervals, while the generated snapshots would actually be
identical without any added benefit.

The backup for an interval is now considered only if is the shortest
interval or there is an oldest snapshot in the previous interval to
copy from for the longer interval.

This means, for example that if 'yearly' is in configuration a new
criterion for qualification for backing up the interval is if
'monthly.11' exists.

The shortest configured interval is always backed up.

Signed-off-by: Eddy Petrișor <eddy.petrisor@gmail.com>