bugfix: if interval has 0 retained, it might need backup
When the check for retained copies was introduced, only shortest and
already populated intervals were thought of.

The correct aproach would have been to consider all unpopulated
intervals and decide later if we can copy from the previous
interval.

Some particular cases should be generalized.

And, finally fixes #2 and fixes #3.

Signed-off-by: Eddy Petrișor <eddy.petrisor@gmail.com>